[Network page]Step 1: Added half of the network functions in a separate network.js file and imported them into index.js Status: Everything Works!
